test: move exception classes from models.py to exceptions.py
